Bug fix: Yesterday I made a change to a test script to use output of          the example program. The change exposed a small error how the          result was printed: since %d was used to print hsize_t value          0 was printed on a big-endian machine instead of a small integer          value. Fixed.
Platforms tested: linew and jam

+++ 1_8/C/H5G/h5ex_g_compact.c
@@ -62,7 +62,7 @@ main (void)
      * Obtain and print the file size.
      */
     status = H5Fget_filesize (file, &size);
-    printf ("File size for %s is: %d bytes\n\n", FILE1, size);
+    printf ("File size for %s is: %d bytes\n\n", FILE1, (int)size);
 
     /*
      * Close FILE1.
@@ -112,7 +112,7 @@ main (void)
      * Obtain and print the file size.
      */
     status = H5Fget_filesize (file, &size);
-    printf ("File size for %s is: %d bytes\n", FILE2, size);
+    printf ("File size for %s is: %d bytes\n", FILE2, (int)size);
     printf ("\n");
 
     /*
